tests/wpt/web-platform-tests/html/webappapis/system-state-and-capabilities/the-navigator-object/navigator_user_agent.tentative.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/webrtc/dictionary-helper.js"></script>
<script>
  test(t => {
    assert_true("getUserAgent" in navigator);
  }, "navigator.getUserAgent() is exposed.");

  promise_test(t => {
    return navigator.getUserAgent()
      .then(ua => {
        // TODO(web-platform-tests/wpt#9106): Use `idlharness.js` once it supports dictionaries.
        assert_string_field(ua, "brand");
        assert_string_field(ua, "version");
        assert_string_field(ua, "platform");
        assert_string_field(ua, "architecture");
        assert_string_field(ua, "model");
      });
  }, "navigator.getUserAgent() returns a UserAgentMetadata object.");
</script>
